Automatic UDF definition from python code
An Algorithm superclass is defined which all algirithms should subclass.
The superclass is equiped with a metaclass whose job is to register all
algorithm methods as MonetDB UDFs. This is a prototype and for the
moment "register" means just append to a list. Later these UDF
definitions will be executed on the DB.