Add ArrayBundle
An ArrayBundle is a class that mimics a two-dimensional numpy array.
However, it does not contain a contiguous chunk of memory. Instead,
every column is contiguous. When provided with a list of arrays, it
constructs an object implementing part of the numpy array api where the
arrays behave as columns of a two-dimensional array. All operations are
carefully implemented as to avoid unnecessary memory copies.

In order to implement all possible type combinations for binary
operations between ArrayBundles, numpy arrays and plain numbers, a
multiple dispatch mechanism is implemented along the lines of the
singledispatch decorator in python's functools.