Add Tensor tables as io udf parameters
UDF can now accept two types of table. Relational tables, the usual
kind, i.e. a set of tuples of heterogeneous fields, and Tensor tables,
i.e. tables encoding vector space objects of any dimension. The
distinction is important as row order doesn't matter in the first case
but is crucial in the second. Moreover, from a performance point of
view, it makes more sense to use contiguous blocks of memory for
tensors, given the kind of operations we are interested in, whereas in
the relational case a simple ensemble of contiguous columns is
sufficient. This leads to to following implementation decision:
On one hand relational tables are a collection of one-dimensional numpy
arrays. On the other hand tensors are unraveled into one long
one-dimensional array, plus some same length int arrays that hold index
information. The number of additional arrays is equal to the number of
dimensions of the tensor. That way a) the whole tensor is always memory
contiguous and b) the original row order can always be recovered from
the index arrays.