Add separate func_name/udf_name params to generate_udf
The idea is that what constitutes a unique name is different in the
python process and in the DB. In python, because the dynamic number of
columns of table parameters, one name is enough. In the DB, where
multiple experiments can run concurrently and where the same function
needs multiple definitions depending on the number of input columns,
the same function name must be further annotated with some execution
context.