Remove dataless classes from UDFCodeGenerator
Output type is not infered anymore but annotated on the udf using a
complex type system where elements of table schemata are seen as type
variables in generic TableT types. The user now can refer to some input datatype,
schema, ncolumns, nrows etc on the return annotation. In the future,
some of these elements could be new (not refering to existing ones in
the input arguments) thus triggering inference.