Working (but dirty) version of udf generation/call mechanism
In this version the idea of inferring the output type is abandoned, for
now. Instead, return types are annotated using TypeVars that reference
elements of the input types.

Another important change is that the udf generator/calling machinery is
now completely agnostic of the names of objects (UDFs, Tables, ...) in
the DB. Instead, it outputs string templates with placeholders for the
node to substitute.