Rewrite udf generator in functional style
UDF generator is completely redesigned from scratch for better
testability, extensibility and clarity. To the degree possible, only
pure functions are used to implement the UDF generator logic. Objects
are mostly used as data structures. An exception to this are the objects
implementing the abstract class ASTNode. These objects represent nodes
in the SQL AST of both UDF definitions and SELECT statements.  All
subclasses of ASTNode implement a single polymorphic method 'compile'
together with a few private/magic helper methods. Another exception
comes from the UDF type system. All input/output types used in UDFs are
subclasses of the abstract class IOType. They are mostly used as data
structures with a few methods useful during the translation/generation
process.

Tensor operations written as pure SQL queries are now implemented in a
very general and modular way. No hard-coded SQL code.

This commit implements features described in multiple Jira tickets:

MIP-4 A new feature is implemented thanks to which a UDF is able to
return a traceback as a string, if an exception is raised.

MIP-43 Implementation of the UDF generator was done using TDD, so the
test coverage is close to 100%.

MIP-5 Previously, a single CREATE TABLE query was used to create a new
table and insert data from the result of a UDF. Now this two queries are
split.

Moreover, some additions in other modules:

1. datatypes.py There is a need for handling the same data types in
different contexts. E.g. in python, SQL or symbolically in DTOs. A new
module is added implementing an enum DType, which holds a single
reference to the various data types present throughout the codebase.
This object exposes a few helper methods for converting between
different representations.

2. udfio.py New additions in the udfio library, used inside MonetDB
containers. Also, unit tests are added covering the additions.

3. Updates in Logistic Regression to fit with the new generator.

4. Small fix in Algorithmexecutor, remove str coercion in literals.