Fix in Linear Regression: move n_obs to compute_summary
`n_obs` was computed in the `fit` method. This gave correct results for
standalone linear regression, where the same matrices are used in both
`fit` and `predict`. However, when performing cross validation this was
no longer correct as we need the value of `n_obs` for the testing
dataset and not the training one.