udfio: remove unused function as_relational_table
This function was wrong and never used. It was supposed to format UDF
outputs when the output type is a relational table. However, this is not
necessary since MonetDB takes care of converting the valid types (list,
dictionary, a numpy array, a numpy masked array or a pandas data frame)
and throws an appropriate message when the output is invalid.

---
https://team-1617704806227.atlassian.net/browse/MIP-536