testcase_generator: allow subclasses to skip input if it makes no sense
The testcase_generator reads the algorithm specs and generates random
inputs, for which the subclasses need to implement the corresponding
output computation based on a state-of-the-art implementation. However,
some random inputs might make no sense for the given algorithm in a way
that the testcase_generator cannot know. With this commit the
`compute_expected_output` method is allowed to return None, in which
case the input is skipped.