udfgenerator: new feature, deferred output schema
One of the main difficulties in the generation of UDF code is that often
the output schema is only known at runtime. This makes static
declaration impossible.

This has lead to the development of a complex output type inference
mechanism in the udfgenerator. This not only adds complexity but it is
also insufficient, as it is not able to cover every use case.

One alternative way to solve this problem is to allow the algorithm flow
to dictate the output schema of some UDFs at runtime. This can be done
by declaring the output schema to be DEFERRED and then pass the actual
schema at runtime as an extra argument in run_udfs_in_local_nodes or
run_udf_in_global_nodes. The udfgenerator then receives the output
schema and generates the appropriate code. This will possibly result in
a few more calls to get_table_schema in the algorithm flow but will lead
to a simpler, more universal and more maintainable solution.

This commit implements the above mechanism by adding a separate
parameter `output_schema` in the chain of functions responsible for
running a UDF on local/global nodes. This parameter should be provided
by the algorithm developer on calls to
`run_udfs_in_local_nodes`/`run_udf_in_global_nodes` for UDFs where the
return type is declared using `relation(schema=DEFERRED)`.

Maybe in the future we should get rid of the type inference mechanism in
favor of the deferred output schema mechanism, and possibly refactor the
algorithms that rely on it.

---
https://team-1617704806227.atlassian.net/browse/MIP-610