merge controller and node exceptions into shared module
The controller exception `BadUserInput` is now used by algorithms for
cases when the provided data is not compatible with the algorithm. This
creates some import problems in production tests due to the fact that
some controller code is missing from the nodes.

renamed:    mipengine/node_exceptions.py -> mipengine/exceptions.py