Algorithm executor: share_outputs is sent to node
When a table is shared between local and global nodes there must be a
column with the node_id in order to discern results from different nodes

When a table is shared between local and global nodes there must be a
column with the node_id in order to discern results from different
nodes. Until now, this column was always added implicitly by having the
UDF generator include the appropriate code in the queries. This lead to
a problem where the UDF generator was attempting to add the node_id
column more than once, if multiple UDFs were chained in a local node.

This commit adds a new mechanism in which the algorithm executor sends
the `share_outputs` list of bools to the node, where it is used by the
UDF generator to append the node_id column only to those tables which
will be shared. That way, when there is a chain of UDFs in the same
node, only the last result will be shared and thus only the last table
will have a node_id column.