Fix algorithms after removing node_id column from results
The previous commit removed a feature from the UDF generator where a
`node_id` column was appended to all UDF results. This introduced a bug
to all algorithms which relied on explicitly ignoring this column in the
result they got from the last UDF. This commit extracts a common
function in the mipengine/algorithms/helpers.py module and fixes the
problem.