Fix algorithms after removing node_id from final result
Commit 3ebb4f83275b5079f4a4e3f8e7d36605fcc79418 introduced a new feature
where the `node_id` column is added only when the corresponding table is
shared. This introduced a bug to all algorithms which relied on
explicitly ignoring this column in the result they got from the last
UDF. This commit extracts a common function in the
mipengine/algorithms/helpers.py module and fixes the problem.