udfgenerator: node_id column is not appended to UDF results
The UDF generator is able to generate code for table returning UDFs
where the resulting table has multiple rows. If these tables are shared
from Local to Global nodes, all the rows end up in the same merge table,
thus there is the need to be able to discern between different node
results. For this reason, the UDF generator is appending a `node_id`
column to every result.

This is however no longer necessary since all algorithms adopted the
transfer object approach, where all UDF results are JSON serialized and
fit in a single row. There is currently no need to append a `node_id`
column since local results occupy each one row in the global node's
merge table.

Moreover, appending the `node_id` to every result lead to a problem
during the implementation of Logistic Regression with cross-validation,
two transformations need to be chained and applied to a local table. The
problem was that the UDF generator was attempting to append the
`node_id` column twice.

It was deemed therefore unnecessary to have the UDF gen appending
`node_id` columns to every result and this feature is completely removed
by the current commit.

Some SMPC tests broke as a result and I had them skipped. I created a
ticket for fixing and un-skipping these tests
https://team-1617704806227.atlassian.net/browse/MIP-683