sql guard: fix literal validation
Due to a MonetDB weakness, the sequence "};" must be prohibited from
literal strings.

So far the literal validator prohibited each of the two characters "}"
or ";" but this is problematic because there are variables in the DB
with ";" in their enumerations.

The current commit changes to validator to disallow the sequence "};",
not the characters contained in it. This still prevents SQL injections
exploiting the above mentioned weakness.