udfgenerator: add placeholder input type
The new placeholder type allows to declare, in the UDF definition, an
arbitrary placeholder assigment of the form

    x = $some_name

This allows passing values inside the UDF definition which cannot be
known to the UDF generator but which are known to the Node. The Node
will have then to replace the placeholder with the appropriate value.