refactor udfgenerator: split smpc workflow
Currently the UDF generator is aware of the SMPC mechanism. This is not
necessary and the two mechanisms should be decoupled. Moreover, the
workflow when the SMPC mechanism is on is intermingled with the workflow
when it's off.

This commit is a first step towards decoupling the two mechanisms. It
doesn't actually achieve the decoupling, but tries to separate the two
workflows to some degree.

All classes involving SMPC are moved to a separate module smpc.py. Some
new classes are created, subclassing existing ones from the non-SMPC
case.

However, the part of the SMPC workflow not belonging to classes but to
simple functions is left almost as is, since subclassing is not an
option.