Added fault tolerance on the table creations and deletions.
All the tables that are now created or deleted, if the connection fails and the query succeeds, can be retried without a 'Table already exists' problem.
In the 'monetdb_interface' moved some global variables inside the methods to mock properly.
Removed soft memory limit from docker containers.
Added unit and integration tests testing the 'IF NOT EXISTS' or 'IF EXISTS' clause in the table queries.