Updated all project packages.
Pinned numpy, scipy, pandas, scikit-learn and statsmodels versions to latest, the same as in the monetdb Dockerfile. Added comments as well to avoid difference.
Updated black and isort packages, in pyproject.toml as well as in github actions.
dnspython was pinned to 2.2.1 due to a bug in eventlet https://github.com/eventlet/eventlet/issues/781 .