udfgen fix: use qualified method names to avoid collisions
When a new function is stored in the UDF registry, a unique name is
generated based on the function and module names. This resulted in name
collisions, since one module could host different classes with the same
method names.

This is now fixed by using the qualified method name instead of the base
name.