udf generator: add mechanism for ad hoc UDFs
The mechanism for ad hoc UDFs is offered by an abstract class
AdhocUdfGenerator. Users implementing it can define arbitrary queries
for running a UDF, bypassing the need to define a python function first.
This allows, among other things, for pure SQL UDFs.

To unify this with the existing UDF mechanism, i.e. via a python function,
a factory class UdfGenerator is implemented. When UdfGenerator is
instantiated, the returned object is either a AdhocUdfGenerator or a
PyUdfGenerator (the old UdfGenerator).

---
https://team-1617704806227.atlassian.net/browse/MIP-757