KFold: add deterministic property based tests
k-fold splitting is a source of non-determinism due to it being
sensitive to row order, which is not guaranteed by the DB. This leads to
non-deterministic tests which fail occasionally. This commit adds
property based tests which verify that the splitting is correct in a
deterministic way. Moreover, the KFold class is tested in isolation, not
relying on specific algorithms.