NODE service get_table_data task is only allowed for published tables.
https://team-1617704806227.atlassian.net/browse/MIP-825
The 'get_table_data' task uses the db public user to fetch data so it only works for published data.
Added 'use_public_user' parameter in the monetdb_facade method, so that the public user is used for the queries.
Added 'get_table_data_from_db' method in the standalone tests (conftest.py) and changed all tests to use this method instead of calling the NODE task.
Refactored many standalone tests to use ONLY the task they are testing, instead of using multiple NODE tasks to ensure the proper result. Instead direct DB queries are made.
Added on/off switch named 'protect_local_data'.
When the switch is on, the 'get_table_data' will use the public user, otherwise not.
The GLOBALNODE should always have this variable switched off.
In the LOCALNODE this variable should be off ONLY in testing scenarios.