Node and Algorithms packages restructuring.
https://team-1617704806227.atlassian.net/browse/MIP-841

In order to incorporate the future development of native-python execution there were some packages restructuring.
The algorithms are now split in 'in_database' and 'native_python' packages.
The 'udfgen' is now part of the 'in_database' package of the 'algorithms' package.

Node service changes:
- Celery tasks were separated from the application logic.
- The application logic was added in a 'services' package.
- Merge/remote tables methods were moved inside the 'tables' package.
- Node_info and 'cleanup' methods were separated in different packages.

Some minor packages restructuring and renaming in the exareme2 parent package.