feat: Controller execution refactoring
The controller class will only contain non-algorithm specific components, sych as the WLA, the cleaner and the task handlers.
The algorithm strategy was enhanced to contain the controller and any algorithm specific information needed, such as the algorithm dto.
A new factory method was added to be used in the api to start the algorithm. The factory chooses the proper controller and then the algorithm strategy passing the controller as a parameter of the strategy.
Removed the algorithm type completely from the algorithm request.